refactor: Integer -> Long

+++ part5/pass-web/src/main/java/com/fastcampus/pass/repository/packaze/PackageEntity.java
@@ -14,7 +14,7 @@ import lombok.ToString;
 public class PackageEntity extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // 기본 키 생성을 DB에 위임합니다. (AUTO_INCREMENT)
-    private Integer packageSeq;
+    private Long packageSeq;
 
     private String packageName;
     private Integer count;
